[Frontend] Consolidate entrypoint exception handler (#52261)
