+++ tests/entrypoints/cohere/test_api_router.py
@@ -30,8 +30,10 @@ from vllm.entrypoints.cohere.protocol import (
     CohereChatV2Response,
 )
 from vllm.entrypoints.openai.engine.protocol import ErrorInfo, ErrorResponse
-from vllm.entrypoints.serve.utils.server_utils import (
+from vllm.entrypoints.serve.exception_handling.handlers.http import (
     http_exception_handler,
+)
+from vllm.entrypoints.serve.exception_handling.handlers.validation import (
     validation_exception_handler,
 )
