+++ tests/entrypoints/serve/exception_handling/test_http_status_metrics.py
