+++ vllm/entrypoints/serve/engine/serving.py
@@ -12,8 +12,8 @@ from vllm.entrypoints.openai.models.serving import (
     OpenAIServingModels,
 )
 from vllm.entrypoints.pooling.typing import AnyPoolingRequest
+from vllm.entrypoints.serve import create_error_response
 from vllm.entrypoints.serve.engine.typing import AnyRequest
-from vllm.entrypoints.serve.utils.error_response import create_error_response
 from vllm.entrypoints.serve.utils.request_logger import RequestLogger
 from vllm.exceptions import VLLMNotFoundError
 from vllm.inputs import EngineInput
