+++ vllm/entrypoints/serve/utils/api_utils.py
@@ -10,7 +10,6 @@ from logging import Logger
 from string import Template
 from typing import Any
 
-import regex as re
 from fastapi import Request
 from fastapi.exceptions import RequestValidationError
 from fastapi.responses import JSONResponse, StreamingResponse
@@ -307,17 +306,6 @@ def process_lora_modules(
         else:
             lora_modules += default_mm_lora_paths
     return lora_modules
-
-
-def sanitize_message(message: str) -> str:
-    """Strip memory addresses, tracebacks, and file paths from error messages."""
-    message = re.sub(r" at 0x[0-9a-f]+>", ">", message)
-    message = re.sub(r'\n?\s*File "[^"]+", line \d+, in \S+(\n\s+.*)?', "", message)
-    message = re.sub(
-        r"/(?:home|usr|opt|var|tmp|root|lib|mnt|srv)(?:/[\w.\-]+)+", "<path>", message
-    )
-    message = re.sub(r"(?:/[\w\-]+)+/[\w\-]+\.\w+", "<path>", message)
-    return message.strip()
 
 
 def log_version_and_model(lgr: Logger, version: str, model_name: str) -> None:
