+++ vllm/entrypoints/speech_to_text/realtime/connection.py
@@ -14,7 +14,7 @@ from starlette.websockets import WebSocketDisconnect
 
 from vllm import envs
 from vllm.entrypoints.openai.engine.protocol import ErrorResponse, UsageInfo
-from vllm.entrypoints.serve.utils.api_utils import sanitize_message
+from vllm.entrypoints.serve.exception_handling.utils import sanitize_message
 from vllm.exceptions import VLLMValidationError
 from vllm.logger import init_logger
